refactor(extensions): split commands/extensions.rs into seven submodules
The 1003-LOC commands/extensions.rs spanned seven unrelated OS surfaces
(menu items, native App menu, tray, file-watcher, debounced reload,
project extension watching, npm installer). Split into a directory
+ mod.rs, mirroring the recent commands/fs/ and shell/ layouts.

Pure relocation — no logic changes. Each submodule owns one concern
and its tests live alongside the code they exercise. Cross-module
helpers stay `pub(super)`; only the seven items lib.rs references
are publicly re-exported through `mod.rs` glob re-exports (required
so Tauri's macro-generated `__cmd__*` siblings stay reachable to
`generate_handler!`).

Layout:
  menu_items.rs    — ExtensionMenuItem, ExtensionMenuStore, set_extension_menu_items
  app_menu.rs      — install_app_menu
  tray.rs          — TRAY_ID, install_tray
  watcher.rs       — AgentWatcher, start_agent_watcher
  reload.rs        — DebounceMsg, run_debounce_worker
  project_watch.rs — watch_project_extensions, unwatch_project_extensions
  installer.rs     — validate + output_tail + install_aethon_extension

Test migration: the three source-grep regression tests now read their
respective new file names (`tray.rs`, `app_menu.rs`, `menu_items.rs`)
via include_str!. All 4 original cases pass under their new module
paths.

Verified: cargo check + clippy -D warnings + cargo test --lib (158
passing) + nix develop -c check (1124 vitest cases) + treefmt.
